requirements.txt
Flask==3.0.3
matplotlib==3.9.2
openai==1.52.0
pandas==2.2.3
azure-search-documents==11.5.1
azure-storage-blob==12.23.1
